+++ cmd/psweb/main.go
@@ -33,7 +33,7 @@ import (
 
 const (
 	// App version tag
-	version = "v1.4.6"
+	version = "v1.4.7"
 )
 
 type SwapParams struct {
@@ -2040,10 +2040,16 @@ func convertPeersToHTMLTable(peers []*peerswaprpc.PeerSwapPeer, allowlistedPeers
 
 			if stats.FeeSat > 0 {
 				flowText += "\nRevenue: +" + formatWithThousandSeparators(stats.FeeSat)
+				if stats.RoutedOut > 0 {
+					flowText += "\nRevenue PPM: " + formatWithThousandSeparators(stats.FeeSat*1_000_000/stats.RoutedOut)
+				}
 			}
 
 			if stats.PaidCost > 0 {
 				flowText += "\nCosts: -" + formatWithThousandSeparators(stats.PaidCost)
+				if stats.PaidOut > 0 {
+					flowText += "\nCosts PPM: " + formatWithThousandSeparators(stats.PaidCost*1_000_000/stats.PaidOut)
+				}
 			}
 
 			tooltip += flowText