+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Versions
 
+## 1.4.7
+
+- Remove resthost from peerswap.conf
+- Fix panic in appendInvoices
+- Add revenue and costs to tooltips
+
 ## 1.4.6
 
 - Estimate swap fees before submitting